### 2022-12-09 #5000 @editorialbot
[PRE REVIEW]: RCzechia: Spatial Objects of the Czech Republic
**Submitting author:** @jlacko (<a href="http://orcid.org/0000-0002-0375-5156">Jindra Lacko</a>)
**Repository:** https://github.com/jlacko/RCzechia
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.10.0
