### 2022-12-09 4cccff5
"near abroad" in rasters / 25km
