+++ DESCRIPTION
@@ -44,6 +44,6 @@ SystemRequirements:
   GDAL (>= 2.2.3), 
   GEOS (>= 3.6.2),
   PROJ (>= 4.9.3)
-RoxygenNote: 7.2.2
+RoxygenNote: 7.2.3
 VignetteBuilder: knitr
 NeedsCompilation: no
