### 2022-12-09 #5000 @arfon
:wave: @jlacko – thanks for your submission to JOSS. Because this library looks limited to downloading and plotting spatial data, it may not be in scope for us. I'm going to ask our editorial team to give their input – we should have an answer for you in the next week or two.
