### 2022-12-11 #5000 @jlacko
Thank you for your consideration; I will wait for instructions.

Two comments only:
- the package does indeed serve spatial data (plus wraps a geocoding and a reverse geocoding API), plotting is a common and a basic use case - which is why I focused my sample code on it - but far from the only one. Significant effort has been put to curating and aligning the data served.
- a functionally close package by our northern neighbors has been published a while back in JOSS / rgugik https://www.theoj.org/joss-papers/joss.02948/10.21105.joss.02948.pdf which led me to believe RCzechia might be relevant for JOSS as well.

Of course no two packages are exactly alike, and editorial policy evolves, so this may not be the case (anymore).
