### 2022-12-15 ee179d8
updated cran check badge / https://recology.info/2022/06/cran-checks-badges/
