### 2022-12-26 #5000 @arfon
:wave: @martinfleis – on reflection I think this might be a better fit for you than @timtroendle. Would you be willing to edit this submission for us?

@timtroendle – I'll invite you on https://github.com/openjournals/joss-reviews/issues/4974#issuecomment-1350530216 instead 😸
