### 2023-01-16 #5000 @martinfleis
Hi @nickbearman, and @paleolimbot, would you be willing to review this JOSS submission? It is not a large package so it should be quite straightforward from your side. Thank you!
