### 2023-01-16 #5000 @nickbearman
Hi @martinfleis in principle yes, but I have a question.

This submission is an R library that provides spatial data for the Czech Republic. Given that it only provides data, does it count as research software?

From: https://joss.theoj.org/about
Scope & submission requirements

> JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library; or similar.

So does it qualify? I had a quick look in issues, but couldn't see anything relevant. My interpretation would be that it's only data, so it's not software, so doesn't qualify but I don't know if that is consistent with the journal's aims.

But Dyba, K., & Nowosad, J. (2021). rgugik: Search and Retrieve Spatial Data from the Polish
108 Head Office of Geodesy and Cartography in R. Journal of Open Source Software, 6(59),
109 2948. https://doi.org/10.21105/joss.02948 is already published here, which is kind of similar.

Comments welcome :-)

Thanks!
