### 2023-01-16 #5000 @martinfleis
Hi @nickbearman, this is a good point. After the submission the paper has been flagged as possibly out of scope but the editorial team internally agreed that it is wihtin the scope as @arfon mentions in https://github.com/openjournals/joss-reviews/issues/5000#issuecomment-1365199719. In my understanding, the package itself does not contain the data but is a wrapper around various sources to provide an R interface to the data portal, similarly to the package from you linked.

So from this perspective, the JOSS considers it within scope. If that is okay with you, I'll assign you as a reviewer.
