+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: RCzechia
 Type: Package
 Title: Spatial Objects of the Czech Republic
-Version: 1.10.0
-Date: 2022-11-25
+Version: 1.10.1
+Date: 2023-01-17
 Authors@R: 
     person("Jindra", "Lacko", , "jindra.lacko@gmail.com", role = c("aut", "cre"),
     comment = c(ORCID = "0000-0002-0375-5156"))
