+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: RCzechia
 Type: Package
 Title: Spatial Objects of the Czech Republic
 Version: 1.10.1
-Date: 2023-01-17
+Date: 2023-01-20
 Authors@R: 
     person("Jindra", "Lacko", , "jindra.lacko@gmail.com", role = c("aut", "cre"),
     comment = c(ORCID = "0000-0002-0375-5156"))
