+++ joss-paper/paper.md
@@ -1,0 +1,138 @@
+---
+title: 'RCzechia: Spatial Objects of the Czech Republic'
+tags:
+  - R
+  - spatial
+  - vector data
+  - country specific Czechia
+authors:
+  - name: Jindra Lacko 
+    orcid: 0000-0002-0375-5156
+    affiliation: 1
+affiliations:
+ - name: Prague University of Economics and Business
+   index: 1
+date: 1 November 2022
+bibliography: paper.bib
+
+---
+
+# Summary
+`RCzechia` is a R package providing spatial objects and spatial information relevant in the context of the Czech Republic for spatial data analysis and visualization purposes. The package uses `sf` data format to serve the most commonly used administrative areas and natural objects. 
+
+As the underlying data is by necessity larger than CRAN package size limits allow, the data is stored externally and a working internet connection is required to use the package.
+
+# State of the field
+The history of spatial data analysis in `R` is long and respectable @bivand21. The first packages focusing specifically on providing spatial data originate from the `S` days @becker_wilks93, with `maps` @deckmyn22 being one of the oldest packages in continuous use on CRAN (since 2003). The early packages used pattern of storing spatial data internally, which created a hard limit on volume and level of detail stored. 
+
+With the advent of `sp` @pebesma_bivand05 and later `sf` @pebesma18 platforms for handling spatial data the universe of packages focused on providing spatial data blossomed. There are packages with global focus, such as `rnaturalearth` @south17 and regional focus like `giscoR` @hernangomezdiego22 oriented at the EU. Number of packages are country specific, such as `tigris` @walker_rudis22 for the US, or `rgugik` @dyba_nowosad21 for Poland. With current near universal and reliable internet access a new pattern has emerged, with spatial data packages accessing cloud stored data files as required (caching them within the limits set by the [CRAN repository policy](https://cran.r-project.org/web/packages/policies.html)), and distributing only lightweight code.
+
+In the context of Czech Republic there exists `CzechData` package @caha21, with somewhat overlapping functionality but available only on GitHub. The CRAN package `czso` @bouchal22 interfaces API of the Czech Statistical Office [ČSÚ](https://www.czso.cz/csu/czso/home), providing access to statistical data about Czech administrative areas (without the spatial information itself). Package `pragr` @bouchal20, available on GitHub, provides geodata about the city of Prague.
+
+# Statement of need
+No country specific spatial data package has been published on CRAN for the Czech Republic to date, creating a need that could be filled using global or regional packages only to a limited extent.
+
+While there there are open data resources available for researchers, mostly in the format of ESRI Shapefiles, these have a number of practical disadvantages. They have to be located and downloaded individually, and their users in R context face additional hurdles, such as conflicting Coordinate Reference Systems and character encodings. In addition some publicly available datasets are topologically invalid and many are too detailed for use by non GIS specialized audience.
+
+# Features
+The package provides two distinct sets of spatial objects: administrative areas, and natural objects. In addition API interface wrapping is provided for geocoding and reverse geocoding functions.
+
+### Administrative area polygons:
+
+* **republika**: borders of the Czech Republic as a polygon
+* **kraje**: 14 regions (NUTS3 areas) of the Czech Republic + Prague as a special case
+* **okresy**: 76 districts (LAU1 areas) of the Czech Republic + Prague as a special case
[…]
