+++ joss-paper/paper.bib
@@ -1,0 +1,232 @@
+
+@article{korecky_etal21,
+	title = {Genetic diversity of {N}orway spruce ecotypes assessed by {GBS}-derived {SNPs}},
+	volume = {11},
+	doi = {10.1038/s41598-021-02545-z},
+	abstract = {We investigated the genetic structure of three phenotypically distinct ecotypic groups of Norway spruce (Picea abies) belonging to three elevational classes; namely, low- (acuminata), medium- (europaea), and high-elevation (obovata) form, each represented by 150 trees. After rigorous filtering, we used 1916 Genotyping-by-Sequencing generated {SNPs} for analysis. Outputs from three multivariate analysis methods (Bayesian clustering algorithm implemented in {STRUCTURE}, Principal Component Analysis, and the Discriminant Analysis of Principal Components) indicated the presence of a distinct genetic cluster representing the high-elevation ecotypic group. Our findings bring a vital message to forestry practice affirming that artificial transfer of forest reproductive material, especially for stands under harsh climate conditions, should be considered with caution.},
+	pages = {23119},
+	journaltitle = {Scientific Reports},
+	shortjournal = {Scientific Reports},
+	author = {Korecký, Jiři and Čepl, Jaroslav and Stejskal, Jan and Faltinová, Zuzana and Dvořák, Jakub and Lstibůrek, Milan and El-Kassaby, Yousry},
+	date = {2021-11-30},
+	keywords = {rczechia},
+	file = {Korecky et al. - 2021 - Genetic diversity of Norway spruce ecotypes assess.pdf:/home/jindra/Zotero/storage/F4HI929U/Korecky et al. - 2021 - Genetic diversity of Norway spruce ecotypes assess.pdf:application/pdf},
+}
+
+
+@article{brejcha_etal21,
+	title = {Variability of colour pattern and genetic diversity of {S}alamandra salamandra ({C}audata: Salamandridae) in the {Czech Republic}},
+	volume = {70},
+	issn = {2694-7684},
+	url = {https://bioone.org/journals/journal-of-vertebrate-biology/volume-70/issue-2/jvb.21016/Variability-of-colour-pattern-and-genetic-diversity-of-Salamandra-salamandra/10.25225/jvb.21016.full},
+	doi = {10.25225/jvb.21016},
+	shorttitle = {Variability of colour pattern and genetic diversity of Salamandra salamandra (Caudata},
[…]
