+++ DESCRIPTION
@@ -38,8 +38,7 @@ Suggests:
   dplyr,
   lwgeom,
   s2,
-  covr,
-  tidyterra
+  covr
 SystemRequirements: 
   GDAL (>= 2.2.3), 
   GEOS (>= 3.6.2),
