### 2023-01-19 74a795e
close check-tuning (#53)
* correcting winbuilder errors

* regex v buildignore
