### 2023-01-19 jlacko/RCzechia#55 @nickbearman
License: JOSS Review
Hi @jlacko
You don't have a license file at https://github.com/jlacko/RCzechia/blob/master/LICENSE which we need for the JOSS review. https://github.com/openjournals/joss-reviews/issues/5082
Please do update it when you can.
Thanks,
Nick.
