### 2023-01-19 178aaf5
formalize license / relates #55
