### 2023-01-19 jlacko/RCzechia#56 @nickbearman
Data origin: JOSS Review
Just adding things as I go through!

You have a lot of data in here, do you have a comprehensive list of the data, it's source and related information? I think this would we really useful. There is a nice summary in the readme, but you only mention the source for the admin areas.

A table, with name, source, last updated and a link to the source would be really nice. This would a) help people understand where the data come from and b) highlight the usefulness of the package more.

Thanks :-)

https://github.com/openjournals/joss-reviews/issues/5082
