### 2023-01-19 ee2aec2
Libraries, JOSS Review
This is primarily an ease of use thing, but it would be really handy to include the library call at the beginning of your examples.
