### 2023-01-19 jlacko/RCzechia#56 @jlacko
Thanks for the comment, it is definitely a valid one - serving data is the chief purpose of the package.

The approach I took when writing it is that each function serves a {sf} flavored data frame, and the data source + cutoff date (more important for man made objects such as admin areas than say rivers or terrain) is in function documentation / the man pages.

Do you think it this would be a better practice to summarise it somewhere?
