### 2023-01-19 jlacko/RCzechia#58 @nickbearman
Data download times: JOSS Review
A small comment. When pulling in various data sets, some of them are bigger than others. Is it possible to add some type of progress measure, particularly for the bigger data sets?

e.g. when I ran `relief <- vyskopis("actual", cropped = TRUE)`

it said `RCzechia: downloading remote dataset.`

and then just sat there until it had finished. Fortunately it didn't take too long, but if a new user didn't know this they might think something had gone wrong.

https://github.com/openjournals/joss-reviews/issues/5082
