### 2023-01-19 jlacko/RCzechia#56 @nickbearman
Ah - I see. It makes sense for it to be in the help pages, that is really useful. I think it would be worth mentioning it in the readme and/or in one of the examples - if it isn't already.

I think a summary would be useful too. You could do this pragmatically I think. A list of name, title, source would be ideal. You could also include size, but I see this is more work as it's not stored programmatically there I think. Happy for other reviewers to comment too.
