### 2023-01-19 jlacko/RCzechia#59 @nickbearman
Community guidelines:  JOSS Review
You do say people can contribute, but I think it would be good to have a CONTRIBUTING.md file, similar to this:

https://github.com/kadyb/rgugik/blob/master/CONTRIBUTING.md

And a Code of Conduct. These integrate with GitHub's issues etc.

I think it would also be worth mentioning somewhere about adding new data / updating existing data. At the minimum, "If you would like to contribute new data to this library, or notice data that needs updating, please open an issue" or similar.

https://github.com/openjournals/joss-reviews/issues/5082
