### 2023-01-19 jlacko/RCzechia#60 @nickbearman
Install dev version: JOSS Review
I'm not sure if this is a me problem or a code problem.

When I try to install the dev version:

`remotes::install_github("jlacko/RCzechia")`

I get this message:

`Downloading GitHub repo jlacko/RCzechia@HEAD`
then it downloads a bit.
Then:
`Error in utils::download.file(url, path, method = method, quiet = quiet,  :
  download from 'https://api.github.com/repos/jlacko/RCzechia/tarball/HEAD' failed`

Anyone else have this issue?

https://github.com/openjournals/joss-reviews/issues/5082
