### 2023-01-19 jlacko/RCzechia#56 @jlacko
I will give it a think. There _should_ be a way to do it programmatically, as I would rather prefer not to have two manual updates that can get out of sync. Combining it with size makes perfect sense, as some - and the terrain you mentioned in #58 is the worst culprit - datasets are rather large. I have addressed it in the function docs, with a somewhat light hearted manner (size is ... , so proceed with caution and patience). But it may not be immediately obvious, since you had to ask.
Thank you for the point!
