### 2023-01-19 c9571b8
close formalize-license (#61)
* formalize license / relates #55

* license tuning
