### 2023-01-19 jlacko/RCzechia#55 @jlacko
The somewhat underwhelming structure of the LICENSE file is a direct consequence of the CRAN interpretation of the MIT license (= declaring MIT in the license field of the package DESCRIPTION file + having a LICENSE file with a year and a name, full stop). This is on one hand non negotiable requirement for a CRAN listing, and on the other hand not in line with GitHub interpretation of the MIT license terms.
I have kept the CRAN structure as it is (since deviating would be a reason for delisting) but added a LICENSE.md file that is ignored from package build. This pattern is what `usethis::use_mit_license()` promotes, and seems to be the current best practice.
