### 2023-01-21 jlacko/RCzechia#58 @jlacko
Thank you  @nickbearman for the fresh look; I fully agree with your observation.

A progress bar will be reported with each download. I just need to make sure that it does not end up in vignette or some other place where it does not belong.
