+++ NEWS.md
@@ -1,3 +1,9 @@
+## version 1.10.9999 (TBD)
+
+ - MIT license formalized
+ 
+ - Code of Conduct formalized
+
 ## version 1.10.1 (2023-01-20)
 
  - the election districts - volebni_okrsky() - are updated to version as of January 2023 presidential elections
