### 2023-01-22 jlacko/RCzechia#60 @jlacko
In addition to the possible merge issue it seems that the HEAD of the dev version has been unnecessarily big. It contains some backup directories with raw data that are ignored on R package build, but are downloaded via git (and then discarded).
I will move these to a separate branch where they will cause no harm.
