### 2023-01-22 jlacko/RCzechia#60 @jlacko
I have slimmed the git HEAD to ~ 2.4 MB, and the install of dev version should be way smoother now (I have checked).

Thanks for reporting the issue!
