### 2023-01-26 jlacko/RCzechia#58 @nickbearman
Wonderful - this works really well.

Now, when I run:
`relief <- vyskopis("actual", cropped = TRUE)`
I get
`RCzechia: downloading remote dataset.`
and a progress update:
`[50%] Downloaded 72192398 bytes...`

Am guessing this is currently in the DEV version and will propagate to the CRAN version in due course.

Happy for you to close this if you are happy too.
