### 2023-01-26 #5082 @nickbearman
Re License https://github.com/jlacko/RCzechia/issues/55 there seems to be a conflict between CRAN's license requirements and JOSS's license requirements. Personally I would say I am fine with how @jlacko has approached this, but I would also like:
a) a second opinion from another reviewer
and/or
b) a view from @martinfleis
