### 2023-01-26 jlacko/RCzechia#59 @nickbearman
Thanks - I see you have added a link to https://www.contributor-covenant.org/version/2/1/code_of_conduct/ on the README.md page and also added a CODE_OF_CONDUCT.md in the repo. Would it not make more sense to link from README.md to CODE_OF_CONDUCT.md rather than an external site?
