### 2023-01-26 #5082 @martinfleis
> there seems to be a conflict between CRAN's license requirements and JOSS's license requirements

This solution seems to be a standard way of dealing with the issue of having a human-readable License file JOSS requires as well as CRAN version. So I am fine with the solution. Thanks for pointing it out!
