### 2023-01-26 jlacko/RCzechia#59 @jlacko
I confess to having used a bolierplate code of conduct - `usethis::use_code_of_conduct()` - which is  a function that on one hand enforces a certain level of consistency and "best practices", but on the other hand is just that -a bolierplate.

I will double check it, and in addition write a short contributing guide.
