### 2023-01-26 jlacko/RCzechia#58 @jlacko
Thanks for the kind words! I am keeping the issue open because my code had been set to quiet download mode (actively suppressing the progress bar) earlier in the development of the package. I had turned it to "loud" mode = progress is reported, but I no longer remember exactly why it was quiet - so I keep the issue intentionally open to do more tests. Most likely there will be no impact but I like GitHub to remind me to make doubly sure before next CRAN submission (which I plan to make when the review process concludes, in case there will be more changes).
