### 2023-01-29 b4dfc40
CoC link / external >>  internal; relates #59
