### 2023-01-29 67b1e3d
+ contributing guideline / relates #59
