+++ R/silnice.R
@@ -1,12 +1,12 @@
 #' Road Network
 #'
-#' Function returning data frame of roads of the Czech Republic as \code{sf} lines. It has no obligatory parameters.
+#' Function returning data frame of roads of the Czech Republic as `sf` lines. It has no obligatory parameters.
 #'
 #' Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).
 #'
 #' The data is current to December 2020. Downloaded size is 6 MB.
 #'
-#' @return \code{sf} data frame with 59.594 rows of 3 variables + geometry:
+#' @return `sf` data frame with 59.594 rows of 3 variables + geometry:
 #'
 #' \describe{
 #'   \item{TRIDA}{Class of the road: highway = dálnice, speedway = rychlostní silnice, 1st class road = silnice I. třídy, 2nd class road = silnice II. třídy, 3rd class road = silnice III. třídy, other road = neevidovaná silnice}
@@ -14,7 +14,7 @@
 #'   \item{MEZINARODNI_OZNACENI}{International road code}
 #' }
 #'
-#' @source Mapový podklad – Data200, 2021 © Český úřad zeměměřický a katastrální. \url{https://www.cuzk.cz}
+#' @source Mapový podklad – Data200, 2021 © Český úřad zeměměřický a katastrální. <https://www.cuzk.cz>
 #'
 #' @export
 
