### 2023-01-29 03f1f2e
data summary in package docs / relates #56
