+++ R/RCzechia.R
@@ -23,39 +23,28 @@
 #'
 #' All objects are implemented as `sf` data frames.
 #'
-#' @section Administrative regions:
+#' @section Data overview & download sizes:
 #'
-#' | asdf | asdf |
-#' | ---- | ----- |
-#' | ema má mísu | psvz |
-#' | míla je emo | 500 |
-#'
-#'
-#' \itemize{
-#'   \item republika - borders of the Czech Republic
-#'   \item kraje - regions / NUTS3 units
-#'   \item okresy - districts / LAU1 units
-#'   \item orp_polygony - municipalities with extended powers (obce s rozšířenou působností)
-#'   \item obce_polygony - municipalities as polygons
-#'   \item obce_body - municipalities as centroids (points)
-#'   \item casti - city districts (for cities that implement them)
-#' }
-#'
-#' @section Natural objects:
-#' \itemize{
-#'   \item reky - rivers
-#'   \item plochy - water bodies
-#'   \item lesy - woodland areas (more than 30 ha in area)
-#' }
-#'
-#' @section Other objects:
-#' \itemize{
-#'   \item silnice - roads
-#'   \item zeleznice - railroads
-#'   \item KFME_grid - grid cells (faunistické čtverce) according to Kartierung
-#'   der Flora Mitteleuropas methodology
-#'   \item chr_uzemi - protected natural areas (chráněná území)
-#' }
+#' |  **Function call**  |  **Description**  |  **Data Source**  |  **High-Res**  |  **Low-Res** |
+#' | -------- | ------- | ------- | -------------: | ------------: |
+#' | republika | borders of the Czech Republic | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx) | 245.8 KB | internal |
+#' | kraje | 14 regions / NUTS3 units | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx) | 982.2 KB | internal |
+#' | okresy | 76 + 1 districts / LAU1 units | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx) | 2.1 MB | internal |
+#' | orp_polygony | 205 + 1 municipalities with extended powers | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx) | 3.1 MB | *NA* |
+#' | obce_polygony | 6.258 municipalities as polygons | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx) | 13.3 MB | *NA* |
+#' | obce_body | 6.258 municipalities as centroids (points) | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx) | 252.0 KB | *NA* |
+#' | casti | 57 city districts (where available) | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx) | 1.5 MB | *NA* |
+#' | senat_obvody | 81 senate districts | [Czech Statistical Office](https://www.czso.cz/csu/czso/podminky_pro_vyuzivani_a_dalsi_zverejnovani_statistickych_udaju_csu) | 10.5 MB | 50.0 KB |
+#' | volebni_okrsky | 14.733 general election districts | [RÚIAN / ČÚZK](https://www.cuzk.cz/ruian/RUIAN.aspx)  | 75.8 MB | 5.0 MB |
[…]
