### 2023-01-31 bae1b62
close joss-review (#63)
* close formalize-license (#61)

* formalize license / relates #55

* license tuning

* close joss-review (#62)

* ET call home / provide update on download / relates #58

* guidelines clarified / relates #59

* update readme

* enlighten the weight master branch

* no meaningful code is expected on data-backup

* re-animate data-backup as a valid (if somewhat empty) direcotry
