### 2023-01-31 jlacko/RCzechia#59 @jlacko
I have updated README to link properly to both the local code of conduct and the newly introduced contributing guidelines; I am closing the issue.
