+++ man/RCzechia-package.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/RCzechia.R
 \docType{package}
-\name{intro}
-\alias{intro}
+\name{RCzechia-package}
+\alias{RCzechia-package}
 \title{RCzechia: Spatial Objects of the Czech Republic}
 \description{
 A selection of spatial objects relevant to the Czech Republic. Due to CRAN
