+++ man/RCzechia-package.Rd
@@ -55,7 +55,7 @@ All objects are implemented as \code{sf} data frames.
 
 \section{Utility functions}{
 
-In addition three utility functions are implemented to support spatial
+In addition the following utility functions are implemented to support spatial
 workflow:
 
 \itemize{
