### 2023-01-31 f45e531
silly typo (#66)
* re-order help pages / intro back to top

* silly typo
