### 2023-02-17 #5082 @nickbearman
> I have amended the dev version of RCzechia package for the issues raised by @nickbearman. This included update of the journal article (adding library calls to code samples). I am thankful for the comments, as they led to improvement of the package for users.
>
> I am not sure what the usual practice is, but I would like to add @nickbearman as [rev] in package description in a future version, as I really appreciated his view.

Thanks @jlacko that's very kind, I really appreciate that.
