### 2023-02-18 #5082 @paleolimbot
Sorry this review took me a while to get to!

As a quick summary - this package is great! Data access packages are difficult to get right but make life so much easier...I'll note that the work that the author did here is two-fold (if I understand the workflow correctly): you've both provided cleaned up data AND written the R package to access that data (in addition to providing the hosting). The package does indeed have very high test coverage and has been on CRAN for many years (which is no small feat for a package like this one with a lot of dependencies that depends on internet access being available).

I do have some high-level suggestions having maintained a few data access packages in the past. None of these are blockers on accepting this paper.

- At a very high level, many calls to functions in your paper could be replaced by (e.g.) `readr::read_rds("https://rczechia.jla-data.net/"Okresy-R-2021-06.rds")`. I wonder if you can lean into that a bit more (e.g., maybe printing that call or the URL when calling the function)...while calling `RCzechia::okresy()` is nice for interactive use and prototyping, a straight download + `read_rds()` might be more appropriate.
- I think you should strongly consider versioning your data and encourage the version to be passed in as an argument (e.g., message saying `Using okresy(version = "2023-01-01")` if this `version` unspecified). You should have the freedom to update the data when you find a better way to clean it (or when sf updates and invalidates the object that you saved), and users should have a path to a stable dataset (e.g., to avoid surprises when rebuilding a paper that you wrote six months ago).
- Providing a cache would be a nice quality of life improvement. You can see how I did it in argodata ( https://github.com/ArgoCanada/argodata/blob/master/R/zzz.R#L12-L21 )...CRAN basically won't let you specify a sane cache directory (e.g., using rappdirs) in advance so you have to make it opt-in; however, you *can* use a temp directory that's alive only for one R session, which is usually good enough.
- Providing a configurable mirror (see the argodata link above for how I configured that too) would let you (or somebody else) test your package against an updated data source (e.g., set `R_RCZECHIA_MIRROR=file://path/to/pending/change` and run `R CMD check`).

Happy to follow up on any of those...just let me know!
