### 2023-02-21 #5082 @martinfleis
@jlacko I see that from the standpoint of JOSS, nothing stands in the way of moving one step closer towards acceptance. Do you want to make some other changes as per suggestions or do you consider existing version ready?
