### 2023-02-22 #5082 @jlacko
I think it would be the best to leave the article as it stands now. The short term impacts from the review process are already integrated in the development version of the package, and I will push them to CRAN shortly (hopefully with an updated citation). The more longer term impacts will require a more thorough revision of the codebase and it seems impractical to keep the article on hold for much longer.
