### 2023-02-22 #5082 @jlacko
Perfect; I will do that. The current version is on Zenodo as https://zenodo.org/record/7665574

I plan to do some small minor adjustments - update DESCRIPTION and CITATION once the paper is formally published, and push a new major version to CRAN. Will this be OK from your perspective?
