+++ DESCRIPTION
@@ -1,11 +1,13 @@
 Package: RCzechia
 Type: Package
 Title: Spatial Objects of the Czech Republic
-Version: 1.10.0
-Date: 2022-11-25
+Version: 1.11
+Date: 2023-03-01
 Authors@R: 
     person("Jindra", "Lacko", , "jindra.lacko@gmail.com", role = c("aut", "cre"),
-    comment = c(ORCID = "0000-0002-0375-5156"))
+    comment = c(ORCID = "0000-0002-0375-5156")),
+    person("Nick", "Bearman", role = "rev",
+           comment = "Nick reviewed the package for JOSS")
 Maintainer: Jindra Lacko <jindra.lacko@gmail.com>
 Description: Administrative regions and other spatial objects of the Czech Republic.
 URL: https://github.com/jlacko/RCzechia
