+++ R/obce_polygony.R
@@ -1,12 +1,12 @@
 #' Municipalities / communes (obce) as polygons
 #'
-#' Function returning data frame of LAU2 administrative units for the Czech Republic as \code{sf} polygons. It takes no parameters.
+#' Function returning data frame of LAU2 administrative units for the Czech Republic as `sf` polygons. It takes no parameters.
 #'
 #' Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).
 #'
 #' The data is current to June 2021 (i.e changes introduced by act 51/2020 Sb. are reflected). Downloaded size is 13.3 MB (so use with caution, and patience).
 #'
-#' @return \code{sf} data frame with 6.258 rows of 14 variables + geometry
+#' @return `sf` data frame with 6.258 rows of 14 variables + geometry
 #'
 #' \describe{
 #'   \item{KOD_OBEC}{Code of the level I commune (obec).}
@@ -23,7 +23,7 @@
 #'   \item{NAZ_CZNUTS3}{Name of the NUTS3 unit (kraj)}
 #' }
 #'
-#' @source © ČÚZK, 2021 \url{https://vdp.cuzk.cz/}
+#' @source © ČÚZK, 2021 <https://vdp.cuzk.cz/>
 #'
 #' @examples
 #' \donttest{
