+++ R/KFME_grid.R
@@ -6,7 +6,7 @@
 #'
 #' @param resolution Should the function return high or low resolution shapefile? Allowed values are "low" and "high". Default is "low".
 #'
-#' @return \code{sf} data frame with 1092 rows in low resolution and 4368 rows in high resolution
+#' @return `sf` data frame with 1092 rows in low resolution and 4368 rows in high resolution
 #'
 #' \describe{
 #'   \item{ctverec}{KFME code of the grid cell; depending on value of `resolution` parameter either 4 digits, or 4 digits + 1 letter}
