### 2023-02-24 0d7c2bf
close Joss-paper (#67)
* test of joss paper compilation using official example

* repair YAML indentation

* tuning joss paper action

* attempting to fix paper.pdf not uploading

* restoring upload directory known to be working

* out with boilerplate; in with serious work!

* citation included

* redundant figures are redundant

* references need comitting too...

* tuning paragraphs

* asi dobrý pro dnešek...

* 1 picture > 1000 words

* capitalization in references

* no version for SW

* more specific summary

* state of the field extended

* pragr in state of the field

* code included


[…]
