+++ .github/workflows/R-CMD-check.yaml
@@ -6,6 +6,7 @@ on:
     branches-ignore:
       - joss-paper
       - data-backup
+
   # 2) every 4th day of month at CET noon last commit on default branch (i.e. master)
   schedule:
     - cron:  "0 11 4 * *"
