### 2023-02-24 e82b748
names in DESCRIPTION as a vector (#68)
