+++ DESCRIPTION
@@ -3,11 +3,12 @@ Type: Package
 Title: Spatial Objects of the Czech Republic
 Version: 1.11
 Date: 2023-03-01
-Authors@R: 
-    person("Jindra", "Lacko", , "jindra.lacko@gmail.com", role = c("aut", "cre"),
-    comment = c(ORCID = "0000-0002-0375-5156")),
-    person("Nick", "Bearman", role = "rev",
-           comment = "Nick reviewed the package for JOSS")
+Authors@R: c(
+  person("Jindra", "Lacko", , "jindra.lacko@gmail.com", role = c("aut", "cre"),
+  comment = c(ORCID = "0000-0002-0375-5156")),
+  person("Nick", "Bearman", role = "rev",
+         comment = "Nick reviewed the package for JOSS")
+  )
 Maintainer: Jindra Lacko <jindra.lacko@gmail.com>
 Description: Administrative regions and other spatial objects of the Czech Republic.
 URL: https://github.com/jlacko/RCzechia
