### 2023-03-04 #5082 @jlacko
Thank you all for the review and opportunity to learn. The comments from @nickbearman and @paleolimbot  made the package better, and I learned a lot in the process. You rock!
